moved auth code from app.js to correct pages, made fetch vendors function
