+++ fetch-utils.js
@@ -1,6 +1,6 @@
-
 const SUPABASE_URL = 'https://ntegguvcsbrefltqhojv.supabase.co';
-const SUPABASE_KEY = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im50ZWdndXZjc2JyZWZsdHFob2p2Iiwicm9sZSI6ImFub24iLCJpYXQiOjE2NDYwNzIwNjEsImV4cCI6MTk2MTY0ODA2MX0.UsQcV5MVZfD-77jViimoke-OeHyMyc9MLiP18wash6I';
+const SUPABASE_KEY =
+    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Im50ZWdndXZjc2JyZWZsdHFob2p2Iiwicm9sZSI6ImFub24iLCJpYXQiOjE2NDYwNzIwNjEsImV4cCI6MTk2MTY0ODA2MX0.UsQcV5MVZfD-77jViimoke-OeHyMyc9MLiP18wash6I';
 
 const client = supabase.createClient(SUPABASE_URL, SUPABASE_KEY);
 
@@ -18,6 +18,11 @@ export function redirectIfLoggedIn() {
     if (getUser()) {
         location.replace('./other-page');
     }
+}
+
+export async function fetchVendors() {
+    const resp = await client.from('vendors').select();
+    return checkError(resp);
 }
 
 export async function signupUser(email, password) {
@@ -38,6 +43,6 @@ export async function logout() {
     return (window.location.href = '../');
 }
 
-// function checkError({ data, error }) {
-//     return error ? console.error(error) : data;
-// }
+function checkError({ data, error }) {
+    return error ? console.error(error) : data;
+}
