+++ signin/signin.js
@@ -1,0 +1,16 @@
+import { redirectIfLoggedIn, signInUser } from '../fetch-utils.js';
+
+const signInForm = document.getElementById('sign-in');
+const signInEmail = document.getElementById('sign-in-email');
+const signInPassword = document.getElementById('sign-in-password');
+
+signInForm.addEventListener('submit', async (event) => {
+    event.preventDefault();
+    const user = await signInUser(signInEmail.value, signInPassword.value);
+
+    if (user) {
+        redirectIfLoggedIn();
+    } else {
+        console.error(user);
+    }
+});
